Fix deleting proposals
