+++ components/common/PageActions/components/PageActionsMenu.tsx
@@ -96,7 +96,6 @@ export function PageActionsMenu({
           <ArchiveProposalAction proposalId={page.id} containerStyle={{ ml: -2 }} />
         </MenuItem>
       )}
-
       {!hideDuplicateAction && page.type && (
         <DuplicatePageAction
           onComplete={handleClose}
@@ -106,7 +105,6 @@ export function PageActionsMenu({
         />
       )}
       <CopyPageLinkAction path={`/${page.path}`} />
-
       <MenuItem dense onClick={onClickOpenInNewTab}>
         <LaunchIcon fontSize='small' sx={{ mr: 1 }} />
         <ListItemText>Open in new tab</ListItemText>
