+++ components/proposals/ProposalsPageWithProviders.tsx
@@ -2,7 +2,7 @@ import { ProposalDialogProvider } from 'components/proposals/components/Proposal
 import ProposalDialogGlobal from 'components/proposals/components/ProposalDialog/ProposalDialogGlobal';
 import { ProposalsProvider } from 'components/proposals/hooks/useProposals';
 import { ProposalsBoardProvider } from 'components/proposals/hooks/useProposalsBoard';
-import { ProposalsPage } from 'components/proposals/ProposalsPageV2';
+import { ProposalsPage } from 'components/proposals/ProposalsPage';
 import { useFeaturesAndMembers } from 'hooks/useFeaturesAndMemberProfiles';
 import { setTitle } from 'hooks/usePageTitle';
 import { ProposalBlocksProvider } from 'hooks/useProposalBlocks';
